requirements.txt
praw==7.7.1
pandas==2.2.0
numpy==1.26.0
tqdm==4.66.1
nltk==3.8.1
vaderSentiment==3.3.2
yfinance==0.2.36
twelvedata==1.2.20
nasdaqdatalink==1.0.1
requests==2.31.0
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0
dash==2.14.2
python-dotenv==1.0.0